Auth Code Refactored to match Project standards
